requirements.txt
aws-cdk-lib==2.72.1
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.72.1a0
